fix(ai-engine): limpiar bugs y vulnerabilidades de SonarQube — 0 bugs, 0 vulns
- AutomationOrchestrator: quitar @Transactional(readOnly=true) de getQueueStatus (S2229, self-invocation hoy en finishRun)
- CoreApiClient: null-guards en vote/createComment para respuestas vacias de .block() (S2259)
- ActivityProfileService: SQL estatico con filtro de bots via = ANY(?) (array param) en vez de interpolacion (S2077); guards de division y null (S3518/S2259)
- AutomationConfig: quitar password hardcodeada 'Bot@2026' del codigo, binding por yml/env + @NotBlank (S2068)
- CaseGenerator: SecureRandom estatico compartido (S2245/S2119), SHA-256 (S4790), regex sin alternancia ambigua (S5850), backoff 100L*attempt (S2184)
- InteractionExecutor: ThreadLocalRandom en vez de new Random() (S2119)
- EngagementService/AutomationScheduler/AutomationOrchestrator: casts a double/long en aritmetica int (S2184)
- IdentityDataSourceConfig: eliminar double-brace init de DriverManagerDataSource (S3599)
- LiveContextService: quitar volatile de listas (accesos ya bajo lock) (S3077)

+++ services/ai-engine-service/src/main/java/com/etribunal/ai/automation/infrastructure/analytics/EngagementService.java
@@ -48,12 +48,12 @@ public class EngagementService {
     /** Calcula el score ponderado 0-100 a partir de los conteos. Expuesto para tests. */
     public int calculateScore(int votes, int comments, int reactions, int shares, int saves, int views) {
         AutomationConfig.EngagementConfig e = config.getEngagement();
-        double weighted = votes * e.getVotesWeight()
-                + comments * e.getCommentsWeight()
-                + reactions * e.getReactionsWeight()
-                + shares * e.getSharesWeight()
-                + saves * e.getSavesWeight()
-                + views * e.getViewsWeight();
+        double weighted = (double) votes * e.getVotesWeight()
+                + (double) comments * e.getCommentsWeight()
+                + (double) reactions * e.getReactionsWeight()
+                + (double) shares * e.getSharesWeight()
+                + (double) saves * e.getSavesWeight()
+                + (double) views * e.getViewsWeight();
         // Normalización logarítmica para acotar a 0-100 sin que casos virales dominen
         double score = 100.0 * (1.0 - Math.exp(-weighted / 60.0));
         return (int) Math.round(score);
